+++ HTTP/Src/WebSocketClient.cpp
@@ -27,7 +27,7 @@ public:
 	{
 		const char* usedaddr = parseProtocol(buffer, len);
 
-		return usedaddr - buffer;
+		return (uint32_t)(usedaddr - buffer);
 	}
 	void read(String& data) {}
 };
@@ -245,7 +245,7 @@ uint32_t WebSocketClient::sendListSize()
 	Guard locker(internal->manager->sendcontent->mutex);
 	for (std::list<std::string>::iterator iter = internal->manager->sendcontent->sendlist.begin(); iter != internal->manager->sendcontent->sendlist.end(); iter++)
 	{
-		cachesize += (*iter).length();
+		cachesize += (uint32_t)(*iter).length();
 	}
 
 	return cachesize;
